fix: adjustments for PR
