+++ modules/home/src/lib/home/home.component.html
@@ -4,12 +4,4 @@
   [alternativeText]="alternativeText"
 ></lib-about-us>
 <lib-latest-news></lib-latest-news>
-<lib-join-us
-  [joinUs]="{
-    title: 'Filie-se',
-    text: aboutUsText,
-    imgSrc: imgSrc,
-    alternativeText: alternativeText,
-    button: 'Saiba mais'
-  }"
-></lib-join-us>
+<lib-join-us></lib-join-us>
