+++ modules/home/src/lib/join-us/join-us.component.html
@@ -11,8 +11,8 @@
   <img
     [ngSrc]="joinUs.imgSrc"
     [alt]="joinUs.alternativeText"
-    width="389"
-    height="205"
+    [width]="magicNumber.FOUR_HUNDRED"
+    [height]="magicNumber.FOUR_HUNDRED"
     priority
   />
 </section>
