+++ modules/home/src/lib/join-us/join-us.component.scss
@@ -1,6 +1,6 @@
 :host section {
   display: flex;
-  max-width: 75rem;
+  max-width: 76rem;
   gap: 1rem; /* espaçamento entre texto e imagem */
   margin: 0 auto; /* Centraliza a seção horizontalmente */
   padding: 0 1rem;
