+++ modules/home/src/lib/join-us/join-us.constants.ts
@@ -1,0 +1,19 @@
+type joinUs = {
+  title: string;
+  text: string;
+  imgSrc: string;
+  alternativeText: string;
+  button: string;
+};
+
+export const JOIN_US = {
+  title: 'Filie-se',
+  text: 'Junte-se a nós e fortaleça a luta pelos direitos das pessoas com deficiência visual no Pará! Ao se filiar à nossa associação, você contribuirá para ampliar o acesso a oportunidades de inclusão, educação e suporte para a comunidade de cegos. Faça parte dessa rede e ajude a construir um futuro mais acessível para todos.',
+  imgSrc: 'joinus.png',
+  alternativeText: 'Foto de equipe da família ASCEPA',
+  button: 'Saiba mais',
+} as joinUs;
+
+export const MAGIC_NUMBERS = {
+  FOUR_HUNDRED: 400,
+};
